requirements.txt
astroid==1.4.9
bbcode==1.0.24
colorama==0.3.7
dj-database-url==0.4.2
Django==1.10.2
django-ajax==0.2.0
django-ajax-selects==1.5.1
django-annoying==0.10.3
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-contrib-comments==1.8.0
django-crispy-forms==1.6.1
django-djconfig==0.5.3
django-excel==0.0.6
django-fluent-comments==1.4.1
django-haystack==2.5.1
django-infinite-scroll-pagination==0.2.0
django-mailer==1.2.2
django-markdown-deux==1.0.5
django-notification==1.3.3
django-pagedown==0.1.1
django-pagination==1.0.7
django-postman==3.5.1
django-spirit==0.5.0.dev0
django-tag-parser==3.1
django-threadedcomments==1.1
django-user-accounts==1.3.1
djangoajax==2.3.7
et-xmlfile==1.0.1
gunicorn==19.6.0
isort==4.2.5
jdcal==1.3
lazy-object-proxy==1.2.2
lockfile==0.12.2
Markdown==2.6.8
markdown2==2.3.1
mccabe==0.6.1
mistune==0.7.1
odfpy==1.3.4
openpyxl==2.4.5
Pillow==3.1.0
pinax-notifications==3.0.1
psycopg2==2.7.1
pybbm==0.18.4
pyexcel==0.4.5
pyexcel-io==0.3.3
pyexcel-ods==0.3.1
pyexcel-webio==0.0.11
pyexcel-xls==0.3.1
pyexcel-xlsx==0.3.0
pyflakes==1.5.0
pylint==1.6.5
python-akismet==0.3.0
pytz==2016.7
requests==2.13.0
six==1.10.0
sorl-thumbnail==12.3
texttable==0.8.8
unicode-slugify==0.1.4
Unidecode==0.4.20
Whoosh==2.7.0
wrapt==1.10.10
xlrd==1.0.0
xlwt==1.2.0
